Splitted code into more structure
