+++ P_Net_Training_GIoU.py
@@ -18,7 +18,8 @@ from sklearn.preprocessing import LabelEncoder
 import tensorflow as tf
 import tensorflow_addons as tfa
 from custom_giou import GIoU
-from train_utils import train_step, validation_step, make_pnet_confidence_map
+from train_utils import train
+from models import build_pnet_model
 from torch.utils.tensorboard import SummaryWriter
 from tensorflow.keras.layers import *
 from tensorflow.keras.models import Model
@@ -30,19 +31,18 @@ from tensorflow.keras.callbacks import TensorBoard, ModelCheckpoint, EarlyStoppi
 from tensorflow.keras.losses import MeanSquaredError, BinaryCrossentropy, CategoricalCrossentropy
 
 ### Some constants ###
-input_dim = 12 # 48
 # weights_dir = 'road_signs_1'
 # weights_dir = 'road_signs_w_dataloader_1'
 weights_dir = 'road_signs_w_dataloadr_l2norm'
 pnet_tensorboard_logdir = 'pnet_logs'
-rnet_tensorboard_logdir = 'rnet_logs'
-onet_tensorboard_logdir = 'onet_logs'
-
 pnet_weights = f'weights/{weights_dir}/pnet.weights.hdf5'
-rnet_weights = f'weights/{weights_dir}/rnet.weights.hdf5'
-onet_weights = f'weights/{weights_dir}/onet.weights.hdf5'
-
 pnet_configs = f'weights/{weights_dir}/pnet.json'
+input_dim = 12 # 48
+epochs = 100 # 500
+batch_size = 16
+train_dir = "/home/minhhieu/Desktop/Hieu/datasets/GTSRB/outputs/obj/train"
+val_dir = "/home/minhhieu/Desktop/Hieu/datasets/GTSRB/outputs/obj/val"
+test_dir = "/home/minhhieu/Desktop/Hieu/datasets/GTSRB/outputs/test"
 
 if(not os.path.exists(f'weights/{weights_dir}')):
     print('[INFO] Created weight directory ...')
@@ -51,22 +51,7 @@ if(not os.path.exists(f'weights/{weights_dir}')):
 if(os.path.exists(pnet_tensorboard_logdir)):
     print('[INFO] Clearing P-Net log directory ... ')
     shutil.rmtree(pnet_tensorboard_logdir)
-
-if(os.path.exists(rnet_tensorboard_logdir)):
-    print('[INFO] Clearing R-Net log directory ... ')
-    shutil.rmtree(rnet_tensorboard_logdir)
-
-if(os.path.exists(onet_tensorboard_logdir)):
-    print('[INFO] Clearing O-Net log directory ... ')
-    shutil.rmtree(onet_tensorboard_logdir)
     
-epochs = 100 # 500
-batch_size = 16
-
-train_dir = "/home/minhhieu/Desktop/Hieu/datasets/GTSRB/outputs/obj/train"
-val_dir = "/home/minhhieu/Desktop/Hieu/datasets/GTSRB/outputs/obj/val"
-test_dir = "/home/minhhieu/Desktop/Hieu/datasets/GTSRB/outputs/test"
-
 ### Loading dataset ###
 ### Creating the train loader ###
 train_loader = DataLoader(train_dir, format_='darknet',
@@ -80,60 +65,8 @@ val_loader = DataLoader(val_dir, format_='darknet',
 
 train_dataset = train_loader.get_train_dataset()
 val_dataset = val_loader.get_train_dataset()
-
-### Creating the val loader ###
-
-
-### Implement the P-Net architecture ###
-def conv_block(in_filters, out_filters, kernel_size=3, batch_norm=False):
-    inputs = Input(shape=(None, None, in_filters))
-    p_layer = Conv2D(out_filters, kernel_size=kernel_size, strides=(1, 1), padding="valid", kernel_regularizer=l1(2e-4))(inputs)
-    if(batch_norm) : p_layer = BatchNormalization()(p_layer)
-    p_layer = PReLU(shared_axes=[1, 2])(p_layer)
-
-    p_layer = MaxPooling2D(pool_size=(2, 2), strides=(2, 2), padding="same")(p_layer)
-
-    block = Model(inputs = inputs, outputs=p_layer)
-    return block
-
-def build_pnet_model(input_shape=None, batch_norm=True, dropout=False, l2_norm=False, n_classes=2, activation='relu'):
-    if(input_shape is not None):
-        if(input_shape not in [12, 24, 48, 112, 224]):
-            raise Exception('Input shape must be in 12, 24, 48')
-
-    inputs = Input(shape=(None, None, 3))
-    p_layer = conv_block(3, 10, kernel_size=3, batch_norm=batch_norm)(inputs)
-
-    if(input_shape is not None):
-        if(input_shape >= 24):
-            p_layer = conv_block(10, 10, kernel_size=3, batch_norm=batch_norm)(p_layer)
-
-    if(input_shape is not None):
-        if(input_shape >= 48):
-            p_layer = conv_block(10, 10, kernel_size=3, batch_norm=batch_norm)(p_layer)
-
-    if(input_shape is not None):
-        if(input_shape >= 112):
-            p_layer = conv_block(10, 10, kernel_size=3, batch_norm=batch_norm)(p_layer)
-
-    p_layer = Conv2D(16, kernel_size=(3, 3), strides=(1, 1), padding="valid", kernel_regularizer=l1(2e-4))(p_layer)
-    p_layer = PReLU(shared_axes=[1, 2])(p_layer)
-
-    p_layer = Conv2D(32, kernel_size=(3, 3), strides=(1, 1), padding="valid", kernel_regularizer=l1(2e-4))(p_layer)
-    p_layer = PReLU(shared_axes=[1, 2])(p_layer)
-    if(dropout) : p_layer = Dropout(0.5)(p_layer)
-
-
-    if(l2_norm):
-        p_layer = Lambda(lambda x : K.l2_normalize(x, axis=3))(p_layer)
-
-    p_layer_out1 = Conv2D(n_classes, kernel_size=(1, 1), strides=(2, 2), kernel_regularizer=l1(2e-4))(p_layer)
-    # p_layer_out1 = Softmax(axis=3, name='probability')(p_layer_out1)
-    p_layer_out2 = Conv2D(4, kernel_size=(1, 1), strides=(2, 2), activation='sigmoid', name='bbox_regression')(p_layer)
-
-    p_net = Model(inputs, [p_layer_out1, p_layer_out2], name='P-Net')
-
-    return p_net
+steps_per_epoch = train_loader.dataset_len
+validation_steps = val_loader.dataset_len
 
 n_classes = train_loader.n_classes
 configs = {
@@ -143,107 +76,21 @@ configs = {
     'n_classes' : n_classes
 }
 
-### Experimenting with l2 normalization on the final logit layer ###
-pnet = build_pnet_model(input_shape=configs['input_shape'], batch_norm=configs['batch_norm'], dropout=configs['dropout'],
-                        n_classes=configs['n_classes'], l2_norm=True)
 print(f'[INFO] Storing P-Net configuration to {pnet_configs}')
 with open(pnet_configs, 'w') as config_file:
     json.dump(configs, config_file, indent=4, sort_keys=True)
 
+### Experimenting with l2 normalization on the final logit layer ###
+pnet = build_pnet_model(input_shape=configs['input_shape'], batch_norm=configs['batch_norm'], dropout=configs['dropout'],
+                        n_classes=configs['n_classes'], l2_norm=True)
 print(pnet.summary())
 
-### Define training loop and start training ###
-steps_per_epoch = train_loader.dataset_len
-validation_steps = val_loader.dataset_len
-
-def train(model, dataset, val_dataset, weights_file, n_classes=10, steps_per_epoch=1000, validation_steps=100, epochs=100, make_conf_map=False):
-    if(os.path.exists(weights_file)):
-        print('Checkpoint exists, loading to model ... ')
-        model.load_weights(weights_file)
-
-    writer = SummaryWriter(log_dir=pnet_tensorboard_logdir)
-    num_gif_files = 0
-    summary = {'train_bbox_loss' : 0, 'train_cls_loss' : 0, 'train_acc' : 0,
-            'val_bbox_loss': 0, 'val_cls_loss' : 0, 'val_acc' : 0}
-    for i in range(epochs):
-        print(f'Epoch {i+1}/{epochs}')
-        with tqdm.tqdm(total=steps_per_epoch) as pbar:
-            cls_losses = []
-            box_losses = []
-            accuracies = []
-            for j in range(steps_per_epoch):
-                batch = next(iter(dataset))
-
-                cls_loss, bbox_loss, acc = train_step(model, batch, n_classes=n_classes)
-                cls_loss = cls_loss.numpy()
-                bbox_loss = bbox_loss.numpy()
-                acc = acc.numpy()
-
-                cls_losses.append(cls_loss)
-                box_losses.append(bbox_loss)
-                accuracies.append(acc)
-
-                if((j+1) % 100 == 0 and make_conf_map):
-                    num_gif_files += 1
-                    make_pnet_confidence_map(pnet, 'test/test.jpg', num_gif_files)
-
-                pbar.set_postfix({
-                    'cls_loss': f'{np.array(cls_losses).mean():.4f}',
-                    'bbox_loss' : f'{np.array(box_losses).mean():.4f}',
-                    'accuracy' : f'{acc:.4f}'
-                })
-                pbar.update(1)
-
-            summary['train_bbox_loss'] = np.array(box_losses).mean()
-            summary['train_cls_loss'] = np.array(cls_losses).mean()
-            summary['train_acc'] = np.array(accuracies).mean()
-
-        print('Saving model weights ... ')
-        model.save_weights(weights_file)
-        print('Validating ... ')
-        with tqdm.tqdm(total=validation_steps // 5, colour='green') as pbar:
-            cls_losses = []
-            box_losses = []
-            accuracies = []
-
-            for j in range(validation_steps // 5):
-                batch = next(iter(val_dataset))
-                cls_loss, bbox_loss, acc = validation_step(model, batch)
-                cls_losses.append(cls_loss)
-                box_losses.append(bbox_loss)
-                accuracies.append(acc)
-
-                pbar.set_postfix({
-                    'cls_loss' : f'{np.array(cls_losses).mean():.4f}',
-                    'bbox_loss' : f'{np.array(box_losses).mean():.4f}',
-                    'accuracy' : f'{acc:.2f}'
-                })
-                pbar.update(1)
-
-            summary['val_bbox_loss'] = np.array(box_losses).mean()
-            summary['val_cls_loss'] = np.array(cls_losses).mean()
-            summary['val_acc'] = np.array(accuracies).mean()
-       
-        # Log bounding box losses to tensorboard log dir
-        writer.add_scalars('bounding_box_loss', {
-            'train' : summary['train_bbox_loss'],
-            'val' : summary['val_bbox_loss']
-        }, i)
-
-        # Log classification losses to tensorboard log dir
-        writer.add_scalars('classification_loss', {
-            'train' : summary['train_cls_loss'],
-            'val' : summary['val_cls_loss']
-        }, i)
-
-        # Log accuracies to tensorboard log dir
-        writer.add_scalars('accuracy', {
-            'train' : summary['train_acc'],
-            'val' : summary['val_acc']
-        }, i)
-
-        writer.flush()
-
-
-train(pnet, train_dataset, val_dataset, pnet_weights, n_classes=n_classes, steps_per_epoch=steps_per_epoch, validation_steps=validation_steps, epochs=epochs, make_conf_map=True)
+### Start training ###
+train(pnet, train_dataset, val_dataset, pnet_weights, 
+        logdir=pnet_tensorboard_logdir,
+        n_classes=n_classes, 
+        steps_per_epoch=steps_per_epoch, 
+        validation_steps=validation_steps, 
+        epochs=epochs, 
+        make_conf_map=True)
 print('[INFO] Training halted, plotting training history ... ')
