+++ models.py
@@ -1,0 +1,82 @@
+import os
+import cv2
+import json
+import tqdm
+import time
+import shutil
+import random
+import numpy as np
+import matplotlib.pyplot as plt
+
+### Other dependencies ###
+from PIL import Image
+from dataloader.obj_detection import DataLoader
+from sklearn.preprocessing import OneHotEncoder
+from sklearn.preprocessing import LabelEncoder
+
+### Tensorflow dependencies ###
+import tensorflow as tf
+import tensorflow_addons as tfa
+from custom_giou import GIoU
+from train_utils import train
+from torch.utils.tensorboard import SummaryWriter
+from tensorflow.keras.layers import *
+from tensorflow.keras.models import Model
+from tensorflow.keras import backend as K
+from tensorflow.keras.optimizers import Adam
+from tensorflow.keras.metrics import MeanIoU
+from tensorflow.keras.regularizers import l2, l1
+from tensorflow.keras.callbacks import TensorBoard, ModelCheckpoint, EarlyStopping
+from tensorflow.keras.losses import MeanSquaredError, BinaryCrossentropy, CategoricalCrossentropy
+
+### Implement the P-Net architecture ###
+def conv_block(in_filters, out_filters, kernel_size=3, batch_norm=False):
+    inputs = Input(shape=(None, None, in_filters))
+    p_layer = Conv2D(out_filters, kernel_size=kernel_size, strides=(1, 1), padding="valid", kernel_regularizer=l1(2e-4))(inputs)
+    if(batch_norm) : p_layer = BatchNormalization()(p_layer)
+    p_layer = PReLU(shared_axes=[1, 2])(p_layer)
+
+    p_layer = MaxPooling2D(pool_size=(2, 2), strides=(2, 2), padding="same")(p_layer)
+
+    block = Model(inputs = inputs, outputs=p_layer)
+    return block
+
+def build_pnet_model(input_shape=None, batch_norm=True, dropout=False, l2_norm=False, n_classes=2, activation='relu'):
+    if(input_shape is not None):
+        if(input_shape not in [12, 24, 48, 112, 224]):
+            raise Exception('Input shape must be in 12, 24, 48')
+
+    inputs = Input(shape=(None, None, 3))
+    p_layer = conv_block(3, 10, kernel_size=3, batch_norm=batch_norm)(inputs)
+
+    if(input_shape is not None):
+        if(input_shape >= 24):
+            p_layer = conv_block(10, 10, kernel_size=3, batch_norm=batch_norm)(p_layer)
+
+    if(input_shape is not None):
+        if(input_shape >= 48):
+            p_layer = conv_block(10, 10, kernel_size=3, batch_norm=batch_norm)(p_layer)
+
+    if(input_shape is not None):
+        if(input_shape >= 112):
+            p_layer = conv_block(10, 10, kernel_size=3, batch_norm=batch_norm)(p_layer)
+
+    p_layer = Conv2D(16, kernel_size=(3, 3), strides=(1, 1), padding="valid", kernel_regularizer=l1(2e-4))(p_layer)
+    p_layer = PReLU(shared_axes=[1, 2])(p_layer)
+
+    p_layer = Conv2D(32, kernel_size=(3, 3), strides=(1, 1), padding="valid", kernel_regularizer=l1(2e-4))(p_layer)
+    p_layer = PReLU(shared_axes=[1, 2])(p_layer)
+    if(dropout) : p_layer = Dropout(0.5)(p_layer)
+
+
+    if(l2_norm):
+        p_layer = Lambda(lambda x : K.l2_normalize(x, axis=3))(p_layer)
+
+    p_layer_out1 = Conv2D(n_classes, kernel_size=(1, 1), strides=(2, 2), kernel_regularizer=l1(2e-4))(p_layer)
+    # p_layer_out1 = Softmax(axis=3, name='probability')(p_layer_out1)
+    p_layer_out2 = Conv2D(4, kernel_size=(1, 1), strides=(2, 2), activation='sigmoid', name='bbox_regression')(p_layer)
+
+    p_net = Model(inputs, [p_layer_out1, p_layer_out2], name='P-Net')
+
+    return p_net
+
